BlockCache.remove(): detect when removing object that is currently being uploaded, and inform upload thread about that.
Fixes issue 419.


This required some changes to the logic used to keep track of blocks
and objects in transit. Previously, we had an "in_transit" set that
contained both (inode, blockno) tuples and object ids that were
currently in transit. Additionally, a "removed_in_transit" set was
used to keep track of object ids for objects that already became obsolete while
still being in transit.

The problem with this approach was that BlockCache.remove() was
able to tell when deleting a block that was in transit, but it
was unable to add the object id to removed_in_transit in some conditions,
because it only knew about the inode and blockno.

This patches therefore replaces in_transit and removed_in_transit
by two new attributes: in_transit_blocks is a dict of (inode, blockno)
tuples, the the values being the corresponding object ids.
in_transit_objs is a set containing all *non-obsolete* object ids
currently in transit.

This allows remove() to determine the object id for a given
(inode, blockno) tuple and then *remove* it from the in_transit_objs
set to mark it as obsolete.