When upload or removal threads terminate prematurely, BlockCache.destroy() could deadlock when trying to distribute sentinel objects, or crash if not all sentinels are retrieved from the removal queue.
This patch special cases QuitSentinels in the Distributor to avoid the
deadlock, and ignores QuitSentinel in the removal queue when asserting
that the queue is empty.

It also fixes the semantics of Distributor.put() to agree with the
documentation: put() now really waits until another thread has consumed
the object.

Fixes issue #32.