Second attempt to fix issue #32. In addition to deadlocking when trying to enqueue termination sentinel objects for threads that have already terminated (fixed in commit e8394818bd29), we can get a similar problem when trying to upload data if there are no upload threads left.
This changeset addresses the second problem. We introduce wrapper
function to enqueue uploads and removals that do not wait indefinitely
but periodically check if there are actually active consumer threads and
raises an exception if not.

With this framework, we also remove the special casing of QuitSentinel
introduced in commit e8394818bd29 and use the wrapper enqueue functions
instead.

Finally, we fix an issue where a failed upload would still allow
deduplication against the block that failed to upload. Previously,
this resulted in loss of the corresponding cache element.

Fixes issue #32.

--HG--
extra : amend_source : 5ac1b0b127e42c379661d67b49e117e8f2a8a879