ObjectR: make closing idempotent. This fixes a bug where the connection is being reset a second time after it has already been reestablished and is in active use again.
This can happen if an instance of a class 'A' derived from io.IOBase holds a
reference to an ObjectR instance 'B' and is also involved in a reference cycle.
If A.close() unconditionally calls B.close(), then B.close() may be called
twice: once when the A instance is closed, and a second time when it is
garbage collected. This happens when 'A' inherits its destructor from IOBase,
because the IOBase destructor unconditionally calls self.close().

In practice, this was observed to happen for the DecryptFilter and DecompressFilter
classes, but it is not currently clear how the reference cycle is created.

--HG--
extra : histedit_source : 5f64ec6b6e9202fef5e00f0d179837b087829be6%2C3dc4be3bc8b3c796a1b74eb43e7c6dedfe0344de