Make documentation building independent of locale.
When subprocess.Popen is called with universal_newlines, it uses
locale.getpreferredencoding(False) to determine the encoding. However,
this may not agree with the document encoding, in which case
docutils.io.FileInput gives an error.

To fix this problem, we could open the stream in binary mode, or pass
the actual encoding to FileInput. However, an even nicer solution is
to use a temporary file, because this also avoids a lot of code
duplication to implement the various include options.