s3c: Be more lenient about XML namespaces in responses.
Ceph RADOS Gateway (RGW) does not always include XML namespaces on every
response. The Amazon S3 documentation says nothing specific XML namespaces,
and includes examples both with and without namespaces.

If a namespace is omitted in a response, assume that it was the correct
namespace. If a different namespace is provided instead, throw an error.

X-Fixes: https://bitbucket.org/nikratio/s3ql/issues/190/sq3l-should-not-rely-on-xmlns-declarations
Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>
Signed-off-by: Robin Johnson <robbat2@gentoo.org>

--HG--
branch : ceph-rgw