Don't enable warnings in setup.py
There seems to be no point, since no important Python code is executed,
and at the moment it just gives a lot of warnings about deprecated functions
in the standard library (probably used by sphinx).

Warnings were added in commit a844968bb961 for programs in bin/*, and
it looks as if including setup.py was an oversight.