fsck.py(): don't link to objects that haven't been uploaded
At the moment, check_cache() silently assumes that all blocks in the database
actually exist in the backend. Therefore, if cache entry has the same match
as a block in the database, we throw the cache file away and link to the
existing block.

However, if the file system crashes while uploading a block, this block
will already be present in the database. So instead of uploading the data
that's still present in the cache, fsck.s3ql will throw it away and link
to the existing block - that unfortunately refers to a non-existing storage
object.

This commit fixes the problem by dropping all rows in the objects and blocks
tables that refer to objects with negative size, i.e. whose upload was
never completed.

It can still happen that we link against a block that whose associated
object was lost for other reasons. Fixing this would require merging the
checks of the object table and local cache files.