checksum_basic_mapping(): look at contents of bytes objects, not just length
At the moment, we are only including the length of byte and bytearray
objects in the checksum, disregarding the actual contents. In practice,
this means that an attacker can modify the metadata stored with a storage
object in the ComprEnc backend as long as the length remains unchanged.

Fixes issue #272.