Remove obsolete comment about get()/remove() concurrency.
This paragraph has most likely become obsolete with commit d8c3fff00b93.
At the moment, the interaction of get() and remove() is exactly as one
would expect: if get() gets the lock before remove(), the block is
removed - otherwise it remains. The comment muddles the water more
than clarifying the behavior.