Use urllib3 to connect to GCE Metadata server
Dugong does not support timeouts for establishing the TCP
connection (which is the operation that will hang when not running on
GCE). Furthermore, ignoring proxy and SSL settings for connection to
GCE server may actually be a feature (I'd assume that proxy and SSL
settings are typically used for connections outside of the VM
infrastructure).

Fixes: #83.