rmdir patch
I found over time my s3ql_data_ directory had many man empty directories -- like 100,000s of them.  This patch will attempt to remove the storage directory when a block is deleted; if the directory is empty, it deletes; if it's not empty, the os.rmdir returns an error which is harmlessly trapped; if it's empty but a race condition means something was going to put a file in there (.. if that's possible) I set "is_temp_failure" to True, so it'll treat it just like S3 etc. not taking a block and retry it.