Re-work warning configuration
Drop rule to ignore mysterious ImportWarning. This does not seem to be an issue anymore.

Use PYTHONWARNINGS to setup warnings in subprocesses when running unit tests. This means
warnings are enabled at import time already. The filters are refined later on.

Use pytest facilities to handle warnings in the same interpreter. There does not seem to
be a way to fully disable these facilities, so attempts to configure warnings manually all
fail in different ways (cf. https://github.com/pytest-dev/pytest/discussions/10847,
https://github.com/pytest-dev/pytest/discussions/10848).