Drop docstring inheritance
The idea of this feature was to reduce duplication of docstrings when there are multiple
classes that implement the same abstract base class.

Unfortunately, this did not work well in practice because it also the docstrings
invisible when editing the code (including for code analysis tools), and effectively
increased the likelihood for an implementation to go out of sync with its docstring.

The plan is to re-introduce explicit docstrings over time when making changes in affected
functions.