Add versioning for metadata
With the current way that we upload metadata, there is a risk that the remote copy is
corrupted. This can happen either because we are uploading an inconsistent snapshot of the
database (i.e., the database is modified while we are uploading), or because the snapshot
wasn't fully written (e.g. because S3QL crashed).

This means that the only time we can rely on having an intact remote metadata copy is when
the filesystem was cleanly unmounted. This is not good, especially when the file system is
used for backups (and a fatal system fault during the backup would thus also corrupt the
backup).

To address this, we version metadata objects.

Note that with the new setup we still only update the sequence number on unmount. This
means that the most recent, guaranteed intact metadata backup is from the last clean
unmount. This may be a problem for filesystems that remain mounted for a long time.