Make http coroutines compatible with Trio main loop.
We can't use trio.run() for eval_coroutine() since this fails if we're already running in
async context (which may happen, because backend objects are currently used both
by synchronous threads and within the asynchronous main thread).

Therefore, we implement the blocking functions by setting an instance variables
that causes the co_* functions not to block.