s3qlrm: emit delete() rather than invalidate() requests
This means that inotify watchers will now receive delete events.

This also introduces a potential problem in that s3qlrm may now fail with ENOTEMPTY if the
S3QL control file (.__s3ql__ctrl__ ) has been looked up anywhere underneath the directory
that is being removed.

However, the likelihood of this happening is too small to outweigh the value of this. Will
track as issue #319.

Fixes: #316