Remove (undocumented) packaging dependency.
It was used to detect a Swift server that has a special COPY operation (for server side copy). After backend refactorings S3QL does not do explicit copies anymore and cannot use the server side copy optimization. That's why we can just remove the whole version check and detection of this feature.